feat: внедрение системы ретраев и валидации корзины по остаткам
- В Order-service добавлен дефолтный Retry для межсервисного взаимодействия, реагирующий на серверные ошибки
- В Cart-Service реализована проверка наличия товара на складе при добавлении/обновлении.
- В Cart-Service исправлены тесты корзины с учетом новых лимитов количества товаров.
- В Cart-Service изменена логика очистки корзины: теперь удаляются только товары, вошедшие в заказ (по списку IDs). (Для идемпотентности при создания заказа)
- В Cart-Service добавлена обработка исключения размыкателя

+++ cart-service/src/main/java/ru/nvkz/dto/CartItemUpdateDto.java
@@ -1,5 +1,5 @@
 package ru.nvkz.dto;
 
-public record CartItemUpdateDto(Long quantity,
+public record CartItemUpdateDto(Integer quantity,
                                 Boolean isSelected) {
 }
